# openapi.yaml
openapi: 3.0.1
info:
  title: Search Plugin
  description: A plugin that searches for information needed to respond to the user, phrased as {query}.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /search/quick:
    get:
      operationId: quicksearch
      summary: a search optimized for speed of response with a moderate number results.
      parameters:
      - in: query
        name: query
        schema:
            type: string
        required: true
        description: phrase describing the information needed.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'

  /search/full:
    get:
      operationId: fullsearch
      summary: a search optimized for maximum number of results while trying to stay within chatgpt timing. this entrypoint may timeout, and may offer quicker or more complete response if a search request is repeated
      parameters:
      - in: query
        name: query
        schema:
            type: string
        required: true
        description: phrase describing the information needed.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'

components:
  schemas:
    SearchResponse:
      type: object
      properties:
        response:
          type: array
          items:
            $ref: "#/components/schemas/SiteResponse"
          description: list of responses from specific urls visited.
        credibility_definitions:
          type: {'credibility':'description'}
          description: a mapping between '#/components/schemas/SiteResponse credibility values and string definitions for those values.
    SiteResponse:
      type: object
      properties:
        source:
          type: string
          description: the domain name of the page visited, e.g. 'weatherunderground.com'
        url:
          type: string
          description: the full url of the page visited
        credibility:
          type: string
          enum: ['Official Source', 'Whitelisted Source', 'Third-Party Source', 'Blacklisted Source']
          description: source credibility, as defined in 'credibility_definitions'
        text:
          type: string
          description: the text relevant to the user query that was extracted from this url